requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AWS SDK for S3 uploads
boto3>=1.34.0

# HTTP requests for API calls
requests>=2.31.0

# Data validation and serialization
pydantic>=2.5.0

# File upload handling
python-multipart>=0.0.6

# OpenAI API client
openai>=1.3.0

# Environment variable management
python-dotenv>=1.0.0

# Additional utilities
typing-extensions>=4.8.0 